suir archivo

+++ src/controller/Recetas.js
@@ -33,28 +33,47 @@ module.exports = {
         //await db.collection('Receta').insertOne(cultura);
         //await db.collection('cultura').insertMany(cultura);
         //console.log("dato agregado");
+        let rec = {};
 
-        //guarda en cloudinary
-        const result = await cloudinary.v2.uploader.upload(req.file.path);
-        //console.log(result);
-        //res.json("guardado")
-   
-        const rec = {
-            titulo: receta.titulo,
-            contenido: receta.contenido,
-            filename : req.file.filename,
-            path : '/img/uploads/' + req.file.filename,
-            originalname : req.file.originalname,
-            mimetype : req.file.mimetype,
-            size : req.file.size,
-            public_id: result.public_id,
-            url: result.url          
+        if(req.file!=null){
+            //guarda en cloudinary
+            const result = await cloudinary.v2.uploader.upload(req.file.path);  
+            rec = {
+                titulo: receta.titulo,
+                indicaciones: receta.indicaciones,
+                precauciones: receta.precauciones,
+                advertencias: receta.advertencias,
+                otrasEnfermedades: receta.otrasEnfermedades,
+                efectosAdversos: receta.efectosAdversos,
+                modoUso: receta.modoUso,
+                filename : req.file.filename,
+                path : '/img/uploads/' + req.file.filename,
+                originalname : req.file.originalname,
+                mimetype : req.file.mimetype,
+                size : req.file.size,
+                public_id: result.public_id,
+                url: result.url          
+            }
+            //borra de la carpeta public para evitar aumentar espacio
+            await unlink(path.resolve('./src/public' + rec.path));
+        }else{
+            rec = {
+                titulo: receta.titulo,
+                indicaciones: receta.indicaciones,
+                precauciones: receta.precauciones,
+                advertencias: receta.advertencias,
+                otrasEnfermedades: receta.otrasEnfermedades,
+                efectosAdversos: receta.efectosAdversos,
+                modoUso: receta.modoUso,
+                filename : null,
+                path : null,
+                originalname : null,
+                mimetype : null,
+                size : null,
+                public_id: null,
+                url: null          
+            }
         }
-        //console.log(rec)
-        //res.json("recibido")
-
-        //borra de la carpeta public para evitar aumentar espacio
-        await unlink(path.resolve('./src/public' + rec.path));
 
         //guarda en la bd
         await db.collection('Receta').insertOne(rec, (err, result) => {
@@ -77,11 +96,11 @@ module.exports = {
         
         //permite borrarlo de cloudinary
         const receta = await db.collection('Receta').find({_id: ObjectID(dato)}).toArray();
-        //await unlink(path.resolve('./src/public' + receta[0].path));
-        const result = await cloudinary.v2.uploader.destroy(receta[0].public_id)
-        
-        //console.log(result)
-        //console.log(receta)
+
+        if(receta[0].url!=null){
+            //await unlink(path.resolve('./src/public' + receta[0].path));
+            const result = await cloudinary.v2.uploader.destroy(receta[0].public_id)
+        }       
 
         //lo borra de la bd
         await db.collection('Receta').deleteOne({
